chore: condense
